% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{discrete_integ}
\alias{discrete_integ}
\title{Discrete integration}
\usage{
discrete_integ(f, k, xd, x)
}
\arguments{
\item{f}{Function, or vector of function evaluations at \code{c(xd, x)}, the
design points \code{xd} adjoined with the query point(s) \code{x}.}

\item{k}{Order for the discrete integral calculation.}

\item{xd}{Design points. These must be sorted in increasing order.}

\item{x}{Query point(s).}
}
\value{
Discrete integral of \code{f} of order \code{k}, with respect to design
points \code{xd}, evaluated at the query point(s) \code{x}.
}
\description{
Computes the discrete integral of a function (or vector of function
evaluations) of a given order, with respect to given design points, and
evaluated at a given query point.
}
\details{
The discrete integral of a function \eqn{f} of order \eqn{k}, with
respect design points \eqn{x_1 < \ldots < x_n}, and evaluated at a query
point \eqn{x}, is defined as TODO
}
\references{
Tibshirani (2020), "Divided differences, falling factorials, and
discrete splines: Another look at trend filtering and related problems",
Section 3.2.
}
