% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{dot_functions}
\alias{dot_functions}
\alias{.divided_diff}
\title{In place computations}
\usage{
.divided_diff(f, z)
}
\description{
Each "dot" function accepts arguments as in its "non-dot"
counterpart, but peforms computations in place, overwriting the first input
argument (which must be a vector of the appropriate length) with the
desired output.
}
\details{
These functions should not be used unless you are intentionally
doing so for memory considerations and are nonetheless being very careful.

An \strong{important warning:} each "dot" function only works as expected if its
first argument is passed in as a vector of numeric type. If the first
argument is passed in as an integer vector, then since the output must (in
general) be a numeric vector, it cannot be computed in place (Rcpp performs
an implicit cast and copy when it converts this to NumericVector type for
use in C++).

Also, each "dot" function does not perform any error checking on its input
arguments. Use with care. More details on the computations performed by
individual functions are provided below.
}
\section{\code{.divided_diff()}}{

Overwrites \code{f} with all lower-order divided differences: each element \code{f[i]}
becomes the divided difference with respect to centers \code{z[1:i]}.
}

\section{\code{.b_mat_mult()}}{

Overwrites \code{v} with \code{B \%*\% v}, where \code{B} is the extended discrete derivative
matrix as returned by \code{b_mat()}.
}

\section{\code{.h_mat_mult()}}{

Overwrites \code{v} with \code{H \%*\% v}, where \code{H} is the falling factorial basis
matrix as returned by \code{h_mat()}.
}

