% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspline-package.R
\docType{package}
\name{dspline-package}
\alias{dspline}
\alias{dspline-package}
\title{dspline: Tools for Computations with Discrete Splines}
\description{
Discrete splines are a class of univariate piecewise polynomial functions with close connections to divided differences and Newton interpolation. Tools for efficient computations relating to discrete splines are provided here. These tools include discrete differentiation and integration, various matrix computations with discrete derivative or discrete spline bases matrices, and interpolation within discrete spline spaces. Many of these techniques are described in Tibshirani (2020) \href{https://arxiv.org/abs/2003.03886}{arXiv:2003.03886}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/glmgen/dspline}
  \item \url{https://glmgen.github.io/dspline/}
  \item Report bugs at \url{https://github.com/glmgen/dspline/issues}
}

}
\author{
\strong{Maintainer}: Ryan Tibshirani \email{ryantibs@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Addison Hu
}

Other contributors:
\itemize{
  \item Logan Brooks [contributor]
}

}
\keyword{internal}
